+++ CHANGELOG.md
@@ -2,8 +2,14 @@
 
 ### `@liveblocks/react`
 
+- Add new Comments hooks to add/remove reactions.
 - Fix a bug in `useOthers()` that could lead to the warning "The result of
   getServerSnapshot should be cached to avoid an infinite loop"
+
+### `@liveblocks/react-comments`
+
+- Add support for reactions. (👍)
+- Add keyboard navigation to emoji picker.
 
 # v1.4.2